added haerds and token decode

+++ controllers/productController.js
@@ -3,6 +3,9 @@ const productModel = require('../models/productModel')
 module.exports.addProduct = async (req, res) => {
 
     console.log(req.body, "5")
+    if (req.permissions.indexOf('add-product') === -1) {
+        return res.send({ code: 401, message: 'Unauthenticated' })
+    }
 
     const newProduct = new productModel(req.body)
     const isSaved = await newProduct.save()
@@ -15,6 +18,11 @@ module.exports.addProduct = async (req, res) => {
 }
 
 module.exports.getProducts = async (req, res) => {
+
+    if (req.permissions.indexOf('view-products') === -1) {
+        return res.send({ code: 401, message: 'Unauthenticated' })
+    }
+
     const data = await productModel.find({})
     if (data.length > 0) {
         res.send({ code: 200, message: 'Find Success', data: data })
@@ -29,7 +37,9 @@ module.exports.getProducts = async (req, res) => {
 
 module.exports.editProduct = async (req, res) => {
     console.log(req.body, 31)
-
+    if (req.permissions.indexOf('edit-product') === -1) {
+        return res.send({ code: 401, message: 'Unauthenticated' })
+    }
     let newData = {}
 
     if (req.body.name) {
@@ -60,6 +70,9 @@ module.exports.editProduct = async (req, res) => {
 }
 
 module.exports.getProductById = async (req, res) => {
+    if (req.permissions.indexOf('view-product') === -1) {
+        return res.send({ code: 401, message: 'Unauthenticated' })
+    }
     let data = await productModel.findById(req.params.id)
     if (data) {
         res.send({ code: 200, message: 'fetch by id success', data: data })
@@ -71,6 +84,9 @@ module.exports.getProductById = async (req, res) => {
 module.exports.deleteProducts = async (req, res) => {
 
     console.log(req.body, "73")
+    if (req.permissions.indexOf('delete-products') === -1) {
+        return res.send({ code: 401, message: 'Unauthenticated' })
+    }
     const ids = req.body
     const response = await productModel.deleteMany({ _id: { $in: ids } })
     if (response) {
